Rename runnerrouter -> inferencerouter (worker unification)
Routes inference; the worker as a whole hosts more than inference (Hydra
agent desktops too). Pure rename, tests still green.


Spec-Ref: helix-specs@e7957a492:001959_we-need-to-replace-all
