+++ api/pkg/inferencerouter/router_test.go
@@ -1,4 +1,4 @@
-package runnerrouter
+package inferencerouter
 
 import (
 	"errors"
